+++ src/main/java/ch/admin/bj/swiyu/issuer/management/api/exception/RestExceptionHandler.java
@@ -6,10 +6,6 @@
 
 package ch.admin.bj.swiyu.issuer.management.api.exception;
 
-import java.util.stream.Collectors;
-
-import static org.springframework.http.HttpStatus.*;
-
 import ch.admin.bj.swiyu.issuer.management.common.exception.*;
 import lombok.NonNull;
 import lombok.extern.slf4j.Slf4j;
@@ -23,6 +19,11 @@ import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.context.request.WebRequest;
 import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
+
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
+
+import static org.springframework.http.HttpStatus.*;
 
 @Order(Ordered.HIGHEST_PRECEDENCE)
 @ControllerAdvice
@@ -74,11 +75,13 @@ public class RestExceptionHandler extends ResponseEntityExceptionHandler {
                                                                   @NonNull HttpHeaders headers,
                                                                   @NonNull HttpStatusCode status,
                                                                   @NonNull WebRequest request) {
-        var errors = ex.getBindingResult().getFieldErrors().stream()
-                .map(error -> String.format("%s: %s", error.getField(), error.getDefaultMessage()))
-                .sorted()
+        String errors = Stream.concat(
+                        ex.getBindingResult().getFieldErrors()
+                                .stream().map(error -> String.format("%s: %s", error.getField(), error.getDefaultMessage())),
+                        ex.getBindingResult().getGlobalErrors().stream().map(error -> String.format("%s: %s", error.getObjectName(), error.getDefaultMessage()))
+                ).sorted()
                 .collect(Collectors.joining(", "));
-        log.debug("Received bad request. Details: {}", errors);
+
         return new ResponseEntity<>(errors, BAD_REQUEST);
     }
 }