+++ src/main/java/ch/admin/bj/swiyu/issuer/management/domain/credentialoffer/CredentialResponseEncryption.java
@@ -6,15 +6,14 @@
 
 package ch.admin.bj.swiyu.issuer.management.domain.credentialoffer;
 
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
 import java.util.Map;
 
-@Data
-@AllArgsConstructor
-public class CredentialResponseEncryption {
-    private Map<String, Object> jwk;
-    private String alg;
-    private String enc;
+/**
+ * information for encrypting the Credential Response.
+ */
+public record CredentialResponseEncryption(
+        Map<String, Object> jwk,
+        String alg,
+        String enc
+) {
 }